requirements.txt
fastapi==0.104.0
uvicorn==0.24.0
httpx==0.25.0
pydantic==2.5.0
python-dotenv==1.0.0
python-multipart==0.0.6
xai-sdk